+++ components/vf-design-tokens/CHANGELOG.md
@@ -1,3 +1,11 @@
+### 3.0.0
+
+- We're moving the documentation to the `vf-component-library`
+  - https://visual-framework.github.io/vf-core/design-tokens/
+- removes all `.njk` documentation files.
+- hides from Fractal.
+- removes any other files no longer needed.
+
 ### 2.1.0
 
 - update documentation pages for all design tokens used to make use of updated CSS.